Show color for registered tokens
