+++ src/routes/registerToken.tsx
@@ -15,13 +15,15 @@ import { List, Icon } from 'antd';
 
 import autobind from 'autobind-decorator';
 import HexString from '../components/hexString';
+import CopyToClipboard from '../components/copyToClipboard';
 import AppLayout from '../components/appLayout';
 import ExitHandler from '../stores/exitHandler';
 import Network from '../stores/network';
 import Tokens from '../stores/tokens';
 import { CONFIG } from '../config';
+import Token from '../stores/token';
 
-const Item = observer(({ item }) => (
+const Item: React.FC<{ item: Token }> = observer(({ item }) => (
   <List.Item key={item.address}>
     <List.Item.Meta
       title={
@@ -37,9 +39,16 @@ const Item = observer(({ item }) => (
         </Fragment>
       }
       description={
-        <Link to={`/explorer/address/${item.address}`}>
-          <HexString>{item.address}</HexString>
-        </Link>
+        <Fragment>
+          <Link to={`/explorer/address/${item.address}`}>
+            <HexString>{item.address}</HexString>
+          </Link>
+          <br />
+          Color:{' '}
+          <CopyToClipboard copyString={item.color}>
+            {item.color}
+          </CopyToClipboard>
+        </Fragment>
       }
     />
   </List.Item>
@@ -78,22 +87,25 @@ export default class RegisterToken extends React.Component<
   }
 
   render() {
-    const { network, tokens } = this.props;
+    const { tokens } = this.props;
 
     return (
       <AppLayout section="registerToken">
-
         <h1 style={{ marginBottom: 16, marginTop: 32 }}>Registered tokens:</h1>
         <List
           itemLayout="vertical"
           size="small"
+          className="tokens-list"
           dataSource={tokens.ready ? tokens.list : undefined}
           renderItem={item => <Item item={item} />}
         />
 
-        <h1 style={{ marginBottom: 16, marginTop: 32 }}>Propose a new ERC20/ERC721 token:</h1>
+        <h1 style={{ marginBottom: 16, marginTop: 32 }}>
+          Propose a new ERC20/ERC721 token:
+        </h1>
         <div>
-          <Iframe url={CONFIG.tokenFormUrl}
+          <Iframe
+            url={CONFIG.tokenFormUrl}
             position="relative"
             height="997px"
             width="100%"
